% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{bloomberg_colors}
\alias{bloomberg_colors}
\title{Bloomberg Colour Palette}
\format{An object of class \code{character} of length 12.}
\usage{
bloomberg_colors
}
\description{
Colour palette based on the Bloomberg colour template \url{https://www.bloomberg.com/graphics/2018-what-the-midterm-campaign-looks-like-in-your-hometown/}
}
\keyword{datasets}
