% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{bloomberg_cols}
\alias{bloomberg_cols}
\title{Function to extract Bloomberg colors as hex codes}
\usage{
bloomberg_cols(...)
}
\arguments{
\item{...}{Character names of bloomberg_colors}
}
\description{
Function to extract Bloomberg colors as hex codes
}
