% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connect_to_dap}
\alias{connect_to_dap}
\title{Connect to the Red Bull DAP environment}
\usage{
connect_to_dap()
}
\description{
One line connection to the DAP database environment
}
\details{
This function will call on environment variable credentials
to connect to the Red Bull internal DAP database.
}
