% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{redbull_colours}
\alias{redbull_colours}
\title{Red Bull Colour Palette}
\format{An object of class \code{character} of length 7.}
\usage{
redbull_colours
}
\description{
Colour palette to support easy themed plotting for Red Bull projects in R.
}
\keyword{datasets}
