% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{redbull_cols}
\alias{redbull_cols}
\title{Function to extract Red Bull colors as hex codes}
\usage{
redbull_cols(...)
}
\arguments{
\item{...}{Character names of redbull_colours}
}
\description{
Function to extract Red Bull colors as hex codes
}
