% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{redbull_palette_discrete}
\alias{redbull_palette_discrete}
\title{Returns a function which provides a discrete redbull colour palette}
\usage{
redbull_palette_discrete(palette = "redbull", reverse = FALSE, rep = TRUE,
  ...)
}
\arguments{
\item{palette}{Character name of palette from redbull_palettes}

\item{reverse}{Boolean indicating whether palette should be reversed}

\item{rep}{Boolean indicating desired behaviour when there are more levels than available palette options}
}
\description{
Returns a function which provides a discrete redbull colour palette
}
