% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{redbull_palette_interpolate}
\alias{redbull_palette_interpolate}
\title{Return a function which interpolates colours using colorRampPalette
This is used for continuous colour scales}
\usage{
redbull_palette_interpolate(palette = "redbull", reverse = FALSE, ...)
}
\description{
Return a function which interpolates colours using colorRampPalette
This is used for continuous colour scales
}
