% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{redbull_palettes}
\alias{redbull_palettes}
\title{Palette definitions for redbull ggplot2 scales}
\format{An object of class \code{list} of length 7.}
\usage{
redbull_palettes
}
\description{
Palette definitions for redbull ggplot2 scales
}
\keyword{datasets}
