% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{redbull_pals}
\alias{redbull_pals}
\title{Return function to interpolate a Red Bull color palette}
\usage{
redbull_pals(palette = "redbull", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a Red Bull color palette
}
