% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{style_redbull}
\alias{style_redbull}
\title{Add redbull theme to ggplot chart}
\usage{
style_redbull()
}
\description{
This function allows you to add the redbull theme to your ggplot graphics.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
style_redbull()
}
