% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_rb}
\alias{theme_rb}
\title{Add redbull theme to ggplot chart}
\usage{
theme_rb(base_size = 11, base_family = "Helvetica",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\description{
This function allows you to add the redbull theme to your ggplot graphics.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
theme_rb()
}
