% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{redbull_palette_discrete}
\alias{redbull_palette_discrete}
\title{Return function to a Red Bull color palette}
\usage{
redbull_palette_discrete(palette = "redbull", reverse = FALSE,
  rep = TRUE, ...)
}
\arguments{
\item{palette}{Character name of palette in redbull_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to a Red Bull color palette
}
